Raise a warning instead of failing on plugin overwrite
I'm not sure why but my local developer versions of plugins are
registering more than once with different names. I think it might be
because I have a difference between the package name and the module
name. Not sure. Anyways, I'm going to change this for now with a nice
large warning so that I can keep an eye on it and keep developing