Add post-commit hook support
The post-commit hook parses the last commit that happened and turns it
into a dictionary. I'm not super fond of the keys in that dict but I
wanted to stick with something that mirrored what you get from git
directly in the hopes that it would be familiar even if it is
inconsistent.

I reworked the hook installation so that it now installs a pre- and
post-commit hook and uses common code to do so.